feat(edit_module): Make all features that edit module alone and modules

+++ src/stores/data.ts
@@ -13,10 +13,10 @@ interface DataState {
   accounts: Map<string, Account>
   modules: Map<string, Module>
   module: Module
-  modulesToAdd: Map<string, Module>
-  modulesToDelete: Array<string>
+  modulesEdited: Map<string, Module>
   accountsEdited: Map<string, Account>
   tipsEdited: Array<string>
+  moduleEdited: Module
   toSave: boolean
 }
 
@@ -26,10 +26,10 @@ export const useDataStore = defineStore('dataStore', {
     accounts: new Map<string, Account>(),
     modules: new Map<string, Module>(),
     module: {} as Module,
-    modulesToAdd: new Map<string, Module>(),
-    modulesToDelete: Array<string>(),
+    modulesEdited: new Map<string, Module>(),
     accountsEdited: new Map<string, Account>(),
     tipsEdited: Array<string>(),
+    moduleEdited: {} as Module,
     toSave: false,
   }),
   getters: {
@@ -37,10 +37,10 @@ export const useDataStore = defineStore('dataStore', {
     getAccounts: (state: DataState) => state.accounts,
     getModules: (state: DataState) => state.modules,
     getModule: (state: DataState) => state.module,
-    getModulesToAdd: (state: DataState) => state.modulesToAdd,
-    getModulesToDelete: (state: DataState) => state.modulesToDelete,
+    getModulesEdited: (state: DataState) => state.modulesEdited,
     getAccountsEdited: (state: DataState) => state.accountsEdited,
     getTipsEdited: (state: DataState) => state.tipsEdited,
+    getModuleEdited: (state: DataState) => state.moduleEdited,
     needToSave: (state: DataState) => state.toSave,
   },
   actions: {
@@ -50,63 +50,137 @@ export const useDataStore = defineStore('dataStore', {
       this.accounts = new Map<string, Account>();
       this.modules = new Map<string, Module>();
       this.module = {} as Module;
-      this.modulesToAdd = new Map<string, Module>();
-      this.modulesToDelete = Array<string>();
+      this.modulesEdited = new Map<string, Module>();
       this.accountsEdited = new Map<string, Account>();
       this.tipsEdited = Array<string>();
+      this.moduleEdited = {} as Module;
       this.toSave = false;
     },
 
     // ------------------------------ CHANGE LOCAL DATAS ------------------------------
+
+    // --------------------- MODULES
+    checkModulesEdition() {
+      // Check if there is something to save
+      if (!areMapsEqual(this.modules, this.modulesEdited)) {
+        this.toSave = true;
+      } else {
+        this.toSave = false;
+      }
+    },
+    checkModuleEdition() {
+      // Check if there is something to save
+      if (!this.moduleEdited.equals(this.module)) {
+        this.toSave = true;
+      } else {
+        this.toSave = false;
+      }
+    },
     addModule(module: Module) {
+      // Check if title is already used
+      let titleAlreadyUsed = false;
+      for (let moduleCheck of this.modulesEdited.values()) {
+        if (moduleCheck.title === module.title) {
+          titleAlreadyUsed = true;
+        }
+      }
       // Change local datas to add module
-      this.modules.set(module.title, module);
-      this.modulesToAdd.set(module.title, module);
-      // Notice that there is something to save
-      this.toSave = true;
+      if (!titleAlreadyUsed) {
+        this.modulesEdited.set(module.title, module);
+        this.toSave = true;
+      }
     },
     deleteModule(id: string) {
       // Change local datas to delete module
-      if (this.modules.has(id)) {
-        this.modules.delete(id);
-        if (this.modulesToAdd.has(id)) {
-          this.modulesToAdd.delete(id);
-        } else {
-          this.modulesToDelete.push(id);
-        }
+      this.modulesEdited.delete(id);
+      // Update order
+      let order = 0;
+      for (let module of this.getSortedModules()) {
+        order++;
+        module.order = order;
       }
       // Check if there is something to save
-      if (this.modulesToAdd.size == 0 && this.modulesToDelete.length == 0) {
-        this.toSave = false;
-      } else {
+      if (!areMapsEqual(this.modules, this.modulesEdited)) {
         this.toSave = true;
+      } else {
+        this.toSave = false;
+      }
+    },
+    editModule(attribute: string, value: any) {
+      // Change local datas to edit module
+      this.moduleEdited[attribute] = value;
+      // Check if there is something to save
+      this.checkModuleEdition();
+    },
+    increaseModuleOrder(id: string) {
+      // Change local datas to increase module order
+      let module = this.modulesEdited.get(id);
+      if (module && module.order !== 1) {
+        for (let moduleToSwitch of this.modulesEdited.values()) {
+          if (moduleToSwitch.order === module.order - 1) {
+            moduleToSwitch.order++;
+            break;
+          }
+        }
+        module.order--;
+        // Check if there is something to save
+        this.checkModulesEdition();
+      }
+    },
+    decreaseModuleOrder(id: string) {
+      // Change local datas to decrease module order
+      let module = this.modulesEdited.get(id);
+      if (module && module.order !== this.modulesEdited.size) {
+        for (let moduleToSwitch of this.modulesEdited.values()) {
+          if (moduleToSwitch.order === module.order + 1) {
+            moduleToSwitch.order--;
+            break;
+          }
+        }
+        module.order++;
+        // Check if there is something to save
+        this.checkModulesEdition();
+      }
+    },
+    getSortedModules() {
+      // Get sorted modules
+      let sortedModules = Array.from(this.modulesEdited.values());
+      sortedModules.sort((a, b) => a.order - b.order);
+      return sortedModules;
+    },
+
+    // --------------------- TIPS
+    checkTipsEdition() {
+      // Check if there is something to save
+      if (JSON.stringify(this.tips) !== JSON.stringify(this.tipsEdited)) {
+        this.toSave = true;
+      } else {
+        this.toSave = false;
       }
     },
     addTip(tip: string) {
       // Change local datas to add tip
       this.tipsEdited.push(tip);
       // Check if there is something to save
-      if (JSON.stringify(this.tips) !== JSON.stringify(this.tipsEdited)) {
-        this.toSave = true;
-      } else {
-        this.toSave = false;
-      }
+      this.checkTipsEdition();
     },
     deleteTip(index: number) {
       // Change local datas to delete tip
       this.tipsEdited.splice(index, 1);
       // Check if there is something to save
-      if (JSON.stringify(this.tips) !== JSON.stringify(this.tipsEdited)) {
-        this.toSave = true;
-      } else {
-        this.toSave = false;
-      }
+      this.checkTipsEdition();
     },
     editTip(index: number, tip: string) {
       // Change local datas to edit tip
       this.tipsEdited[index] = tip;
       // Check if there is something to save
-      if (JSON.stringify(this.tips) !== JSON.stringify(this.tipsEdited)) {
+      this.checkTipsEdition();
+    },
+
+    // --------------------- ACCOUNTS
+    checkAccountsEdition() {
+      // Check if there is something to save
+      if (!areMapsEqual(this.accounts, this.accountsEdited)) {
         this.toSave = true;
       } else {
         this.toSave = false;
@@ -124,22 +198,14 @@ export const useDataStore = defineStore('dataStore', {
       if (!mailAlreadyUsed) {
         this.accountsEdited.set(account.mail, account);
         // Check if there is something to save
-        if (!areMapsEqual(this.accounts, this.accountsEdited)) {
-          this.toSave = true;
-        } else {
-          this.toSave = false;
-        }
+        this.checkAccountsEdition();
       }
     },
     deleteAccount(id: string) {
       // Change local datas to delete account
       this.accountsEdited.delete(id);
       // Check if there is something to save
-      if (!areMapsEqual(this.accounts, this.accountsEdited)) {
-        this.toSave = true;
-      } else {
-        this.toSave = false;
-      }
+      this.checkAccountsEdition();
     },
     editAccount(id: string, attribute: string, value: any) {
       // Change local datas to edit account
@@ -148,28 +214,33 @@ export const useDataStore = defineStore('dataStore', {
         (account as { [key: string]: any })[attribute] = value;
       }
       // Check if there is something to save
-      if (!areMapsEqual(this.accounts, this.accountsEdited)) {
-        this.toSave = true;
-      } else {
-        this.toSave = false;
-      }
+      this.checkAccountsEdition();
     },
 
     // ------------------------- SEND LOCAL DATAS TO FIREBASE -------------------------
     saveModulesToFirebase: async function () {
       // Add new modules
-      for (let module of this.modulesToAdd.values()) {
-        await addDoc(modulesCollection, module.toJsonObject());
+      for (let [id, module] of this.modulesEdited.entries()) {
+        if (!this.modules.has(id)) {
+          // This is a new module
+          await addDoc(modulesCollection, module.toJsonObject());
+        } else {
+          // This module has been edited
+          await setDoc(doc(modulesCollection, id), module.toJsonObject());
+        }
       }
-      // Delete modules
-      for (let id of this.modulesToDelete) {
-        await deleteDoc(doc(modulesCollection, id));
+      // Delete accounts
+      for (let id of this.modules.keys()) {
+        if (!this.modulesEdited.has(id)) {
+          // This module has been deleted
+          await deleteDoc(doc(modulesCollection, id));
+        }
       }
       this.loadModulesFromFirebase();
     },
     saveModuleToFirebase: async function (id: string) {
       // Replace existing module
-      await setDoc(doc(modulesCollection, id), this.module.toJsonObject());
+      await setDoc(doc(modulesCollection, id), this.moduleEdited.toJsonObject());
       this.loadModuleFromFirebase(id);
     },
     saveTipsToFirebase: async function () {
@@ -246,13 +317,15 @@ export const useDataStore = defineStore('dataStore', {
             modulePartElement = new ModulePartElement(element.image, element.text);
             modulePartElements.push(modulePartElement);
           }
-          modulePart = new ModulePart(part.image, part.order, part.subtitle, modulePartElements);
+          modulePart = new ModulePart(part.image, part.subtitle, modulePartElements);
           moduleParts.push(modulePart);
+
           modulePartElements = Array<ModulePartElement>();
         }
-        let module = new Module(doc.data().icon, doc.data().image, doc.data().title, moduleParts);
+        let module = new Module(doc.id, doc.data().icon, doc.data().image, doc.data().title, doc.data().order, moduleParts);
+        let moduleEdited = module.clone();
         this.modules.set(doc.id, module);
-        moduleParts = Array<ModulePart>();
+        this.modulesEdited.set(doc.id, moduleEdited);
       });
     },
     loadModuleFromFirebase: async function (id: string) {
@@ -274,12 +347,13 @@ export const useDataStore = defineStore('dataStore', {
             modulePartElement = new ModulePartElement(element.image, element.text);
             modulePartElements.push(modulePartElement);
           }
-          modulePart = new ModulePart(part.image, part.order, part.subtitle, modulePartElements);
+          modulePart = new ModulePart(part.image, part.subtitle, modulePartElements);
           moduleParts.push(modulePart);
+
           modulePartElements = Array<ModulePartElement>();
         }
-        this.module = new Module(doc.data().icon, doc.data().image, doc.data().title, moduleParts);
-        moduleParts = Array<ModulePart>();
+        this.module = new Module(doc.id, doc.data().icon, doc.data().image, doc.data().title, doc.data().order, moduleParts);
+        this.moduleEdited = this.module.clone();
       } else {
         // Module does not exist in Firestore collection, redirect on modules page
         router.replace({ name: 'Modules' });
